requirements.txt
llama-index==0.11.4
PyYAML
plotly==5.24.0
streamlit==1.38.0
docx2txt==0.8
python-dotenv==1.0.1
llama-index-vector-stores-chroma==0.3.0